Fix size check for specialised integer-based assignments. (Bug #1572394 - patch provided by Phil Wilson) Backported from 5.4
git-svn-id: file:///home/hardaker/lib/sf-bkups/net-snmp-convert-svnrepo/branches/V5-2-patches@15731 06827809-a52a-0410-b366-d66718629ded

+++ snmplib/check_varbind.c
@@ -75,7 +75,7 @@ netsnmp_check_vb_int_range(const netsnmp_variable_list *var, int low, int high)
     if (NULL == var)
         return SNMP_ERR_GENERR;
     
-    if ((rc = netsnmp_check_vb_type_and_size(var, ASN_INTEGER, sizeof(int))))
+    if ((rc = netsnmp_check_vb_type_and_size(var, ASN_INTEGER, sizeof(long))))
         return rc;
     
     if ((*var->val.integer < low) || (*var->val.integer > high)) {
@@ -93,7 +93,7 @@ netsnmp_check_vb_truthvalue(const netsnmp_variable_list *var)
     if (NULL == var)
         return SNMP_ERR_GENERR;
     
-    if ((rc = netsnmp_check_vb_type_and_size(var, ASN_INTEGER, sizeof(int))))
+    if ((rc = netsnmp_check_vb_type_and_size(var, ASN_INTEGER, sizeof(long))))
         return rc;
     
     return netsnmp_check_vb_int_range(var, 1, 2);
@@ -107,7 +107,7 @@ netsnmp_check_vb_rowstatus_value(const netsnmp_variable_list *var)
     if (NULL == var)
         return SNMP_ERR_GENERR;
     
-    if ((rc = netsnmp_check_vb_type_and_size(var, ASN_INTEGER, sizeof(int))))
+    if ((rc = netsnmp_check_vb_type_and_size(var, ASN_INTEGER, sizeof(long))))
         return rc;
     
     return netsnmp_check_vb_int_range(var, SNMP_ROW_NONEXISTENT,
@@ -136,7 +136,7 @@ netsnmp_check_vb_storagetype(const netsnmp_variable_list *var, int old_value)
     if (NULL == var)
         return SNMP_ERR_GENERR;
     
-    if ((rc = netsnmp_check_vb_type_and_size(var, ASN_INTEGER, sizeof(int))))
+    if ((rc = netsnmp_check_vb_type_and_size(var, ASN_INTEGER, sizeof(long))))
         return rc;
     
     if ((rc = netsnmp_check_vb_int_range(var, SNMP_STORAGE_NONE,
